requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
boto3==1.24.17
botocore==1.27.96
decorator==5.1.1
executing==1.2.0
ipython==8.4.0
jedi==0.18.2
jmespath==1.0.1
joblib==1.2.0
numpy==1.24.3
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post5
SQLAlchemy==1.4.41
stack-data==0.6.2
threadpoolctl==3.1.0
tqdm==4.64.0
traitlets==5.9.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
